requirements.txt
matplotlib~=3.9.2
numpy~=1.26.4
Pillow~=10.4.0
tqdm~=4.66.5
wandb~=0.20.1

torch~=2.7.1+cu126
opencv-python~=4.11.0.86
scipy~=1.13.1
torchvision~=0.22.1+cu126
imageio~=2.33.1